Split python main module into host and base
  Deleted helpers.py
  Starting work toward GH issue #12
