+++ python/asciiserialcom/cli.py
@@ -5,12 +5,12 @@ import logging
 import typer
 import trio
 
-from asciiserialcom.asciiSerialCom import Ascii_Serial_Com
+from asciiserialcom.host import Host
 
 DEFAULT_TIMEOUT = 5
 
 logging.basicConfig(
-    # filename="test_asciiSerialCom.log",
+    # filename="test_hostiiSerialCom.log",
     # level=logging.INFO,
     level=logging.DEBUG,
     format="%(levelname)s L%(lineno)d %(funcName)s: %(message)s",
@@ -23,8 +23,8 @@ async def run_read(timeout, fin_name, fout_name, reg_num):
         async with await trio.open_file(fout_name, "bw") as fout:
             async with await trio.open_file(fin_name, "br") as fin:
                 async with trio.open_nursery() as nursery:
-                    asc = Ascii_Serial_Com(nursery, fin, fout, 8)
-                    result = await asc.read_register(reg_num)
+                    host = Host(nursery, fin, fout, 8)
+                    result = await host.read_register(reg_num)
                     cancel_scope.cancel()
     return result
 
@@ -35,8 +35,8 @@ async def run_write(timeout, fin_name, fout_name, reg_num, reg_val):
         async with await trio.open_file(fout_name, "bw") as fout:
             async with await trio.open_file(fin_name, "br") as fin:
                 async with trio.open_nursery() as nursery:
-                    asc = Ascii_Serial_Com(nursery, fin, fout, 8)
-                    await asc.write_register(reg_num, reg_val)
+                    host = Host(nursery, fin, fout, 8)
+                    await host.write_register(reg_num, reg_val)
                     result = True
                     cancel_scope.cancel()
     return result
@@ -45,7 +45,7 @@ async def run_write(timeout, fin_name, fout_name, reg_num, reg_val):
 async def forward_received_messages_to_print(ch):
     while True:
         msg = await ch.receive()
-        typer.echo(f"{msg.get_packed().decode('ascii')}")
+        typer.echo(f"{msg.get_packed().decode('hostii')}")
 
 
 async def run_stream(timeout, fin_name, fout_name):
@@ -57,12 +57,12 @@ async def run_stream(timeout, fin_name, fout_name):
         async with await trio.open_file(fout_name, "bw") as fout:
             async with await trio.open_file(fin_name, "br") as fin:
                 async with trio.open_nursery() as nursery:
-                    asc = Ascii_Serial_Com(nursery, fin, fout, 8)
+                    host = Host(nursery, fin, fout, 8)
                     nursery.start_soon(forward_received_messages_to_print, recv_ch)
-                    asc.forward_received_s_messages_to(send_ch)
-                    await asc.send_message(b"n", b"")
+                    host.forward_received_s_messages_to(send_ch)
+                    await host.send_message(b"n", b"")
                     await trio.sleep(timeout)
-                    await asc.send_message(b"f", b"")
+                    await host.send_message(b"f", b"")
                     cancel_scope.cancel()
